challenges/igorsgorypasswordsafe/docker/app/requirements.txt
cryptography
Flask==2.2.2
Flask_login==0.6.2
Flask-MySQL==1.5.2
PyMySQL==1.0.2
markupsafe==2.1.2
pycryptodome==3.17
bcrypt==4.0.1